refactor(ui, tauri): completely remove legacy Deep Research feature
Remove the entire Deep Research subsystem to create a clean slate
for the upcoming Council Mode UI (Phase 4).

Frontend (React/TypeScript):
- Delete DeepResearch components (toggle, artifact, 10+ sub-components)
- Delete useDeepResearch hook (research loop, fact extractor, types)
- Delete DeepResearchContext providers (top-level and component-scoped)
- Delete research config constants and useResearchLogs hook
- Delete researchLogger singleton (extract truncatePayload to logging/truncate.ts)
- Clean ChatMessagesPanel: remove ~200 lines of research state, callbacks, JSX
- Clean MessageBubbles: remove ResearchArtifact rendering branch
- Clean message types: remove researchState/isDeepResearch fields
- Clean transport types: remove research metadata from ChatMessageMetadata
- Simplify buildLoadedMessage (no more research branching)
- Remove unused imports (saveMessage, updateMessage, useComposerRuntime)

Tauri/Rust backend:
- Delete src-tauri/src/commands/research_logs.rs (4 commands)
- Remove research_logs from mod.rs and main.rs invoke_handler
- Clean gglib-core chat.rs and gglib-db setup.rs comments

Scripts & docs:
- Update check-tauri-commands.sh allowlists (remove research_logs)
- Update CONTRIBUTING.md, scripts/README.md, gglib-tauri/README.md
- Remove research_logs/ from .gitignore

Tests:
- Delete 2 Deep Research test files (textOnlyStepTracking, typeGuards)

Net: 49 files changed, 85 insertions, 10,316 deletions
