+++ crates/gglib-core/src/domain/chat.rs
@@ -34,7 +34,7 @@ pub struct Message {
     pub role: MessageRole,
     pub content: String,
     pub created_at: String,
-    /// Optional JSON metadata for deep research state, tool usage, etc.
+    /// Optional JSON metadata for tool usage, etc.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub metadata: Option<serde_json::Value>,
 }
@@ -145,7 +145,7 @@ pub struct NewMessage {
     pub conversation_id: i64,
     pub role: MessageRole,
     pub content: String,
-    /// Optional JSON metadata for deep research state, tool usage, etc.
+    /// Optional JSON metadata for tool usage, etc.
     pub metadata: Option<serde_json::Value>,
 }
