+++ src-tauri/src/commands/mod.rs
@@ -3,7 +3,6 @@
 //! After Phase 3 HTTP API consolidation, only OS-specific commands remain:
 //! - llama: Binary installation and status checks
 //! - util: API discovery, menu sync, OS integration
-//! - research_logs: Deep research log file persistence
 //! - app_logs: Frontend-to-backend logging bridge
 //!
 //! All business logic is exposed via HTTP API (gglib-axum).
@@ -11,5 +10,4 @@
 
 pub mod app_logs;
 pub mod llama;
-pub mod research_logs;
 pub mod util;
