--- src/components/DeepResearch/ResearchArtifact/index.ts
@@ -1,2 +1,0 @@
-export { ResearchArtifact } from './ResearchArtifact';
-export type { ResearchArtifactProps } from './types';
