--- src/components/DeepResearch/ResearchArtifact/styles.ts
@@ -1,12 +1,0 @@
-/**
- * Shared class-string constants for ResearchArtifact sub-components.
- *
- * @module components/DeepResearch/ResearchArtifact/styles
- */
-
-/**
- * Base styles for the compact action buttons used in the research plan.
- * Use with `cn(ACTION_BTN_STYLES, ...)` for colour overrides (e.g. "Skip All").
- */
-export const ACTION_BTN_STYLES =
-  'flex items-center gap-1 px-2 py-1 text-[11px] font-medium text-text-secondary bg-background-tertiary border border-border rounded cursor-pointer transition-all duration-150 ease-out hover:text-text hover:bg-background-hover hover:border-[#4a4a4a] disabled:opacity-50 disabled:cursor-default disabled:hover:bg-background-tertiary disabled:hover:border-border';
