--- src/components/DeepResearch/index.ts
@@ -1,14 +1,0 @@
-/**
- * Deep Research Components
- *
- * UI components for the Deep Research mode, including the Research Artifact
- * that displays live research progress and results.
- *
- * @module components/DeepResearch
- */
-
-export { ResearchArtifact } from './ResearchArtifact';
-export type { ResearchArtifactProps } from './ResearchArtifact';
-
-export { DeepResearchToggle } from './DeepResearchToggle';
-export type { DeepResearchToggleProps } from './DeepResearchToggle';
