--- src/hooks/useDeepResearch/index.ts
@@ -1,158 +1,0 @@
-/**
- * Deep Research Mode
- *
- * Implements "Plan-and-Execute" research with structured state management,
- * multi-model routing, and observation injection patterns.
- *
- * @module useDeepResearch
- */
-
-// Core types and state management
-export type {
-  // Configuration
-  ModelEndpoint,
-  ModelRouting,
-  SerializationBudget,
-  // Research plan
-  QuestionStatus,
-  QuestionSource,
-  ResearchQuestion,
-  // Knowledge base
-  FactConfidence,
-  GatheredFact,
-  Contradiction,
-  // Observations
-  PendingObservation,
-  // Search history (multi-round deduplication)
-  SearchRecord,
-  // Round summaries (multi-round context compression)
-  RoundSummary,
-  // Internal tool args
-  AssessProgressArgs,
-  RequestSynthesisArgs,
-  // Core state
-  ResearchPhase,
-  ResearchState,
-  // Human-in-the-loop
-  ResearchIntervention,
-  InterventionRef,
-  // Serialization
-  ResearchContextInjection,
-  SerializedResearchState,
-  // UI
-  ResearchArtifactProgress,
-} from './types';
-
-export {
-  // Factory functions
-  createDefaultRouting,
-  createQuestion,
-  createFact,
-  createInitialState,
-  // Serialization
-  DEFAULT_BUDGET,
-  serializeForPrompt,
-  renderContextForSystemPrompt,
-  serializeState,
-  deserializeState,
-  // State mutations
-  addFacts,
-  updateQuestion,
-  clearObservations,
-  addObservation,
-  advanceStep,
-  setPhase,
-  setError,
-  completeResearch,
-  // Search history helpers
-  isSearchDuplicate,
-  addSearchRecord,
-  linkSearchToFacts,
-  // Round management helpers
-  createRoundSummary,
-  advanceRound,
-  canContinueResearch,
-  getRoundStepBudget,
-  shouldTriggerRoundSoftLanding,
-  // Internal research tools
-  ASSESS_PROGRESS_TOOL,
-  REQUEST_SYNTHESIS_TOOL,
-  INTERNAL_RESEARCH_TOOLS,
-  isInternalResearchTool,
-  MIN_FACTS_FOR_SYNTHESIS,
-  // Validation
-  validateState,
-  // UI helpers
-  getArtifactProgress,
-} from './types';
-
-// Turn message construction
-export type {
-  TurnMessage,
-  TurnMessages,
-  BuildTurnMessagesOptions,
-} from './buildTurnMessages';
-
-export {
-  PHASE_INSTRUCTIONS,
-  buildTurnMessages,
-  buildTurnMessagesWithBudget,
-  estimateTokens,
-  isWithinBudget,
-  shouldIncludeTools,
-  filterResearchTools,
-  getResearchToolsWithInternals,
-  summarizeTurnMessages,
-} from './buildTurnMessages';
-
-// Research loop orchestrator
-export type {
-  ToolDefinition,
-  ToolCall,
-  ToolResult,
-  ToolExecutor,
-  LLMResponse,
-  LLMCaller,
-  RunResearchLoopOptions,
-  ResearchLoopResult,
-} from './runResearchLoop';
-
-export {
-  DEFAULT_MAX_STEPS,
-  SOFT_LANDING_THRESHOLD,
-  MAX_PARALLEL_TOOLS,
-  TOOL_TIMEOUT_MS,
-  CONSECUTIVE_UNPRODUCTIVE_LIMIT,
-  HARD_MAX_STEPS,
-  MAX_TEXT_ONLY_STEPS,
-  MAX_LOOP_ITERATIONS,
-  runResearchLoop,
-  createProxyLLMCaller,
-  createRegistryToolExecutor,
-} from './runResearchLoop';
-
-// Fact extraction
-export type {
-  ExtractionLLMCaller,
-  ExtractFactsOptions,
-  ExtractFactsResult,
-} from './factExtractor';
-
-export {
-  DEDUP_SIMILARITY_THRESHOLD,
-  MAX_FACTS_RETAINED,
-  MAX_FACTS_PER_OBSERVATION,
-  extractFacts,
-  pruneFacts,
-  calculateSimilarity,
-  wouldBeDuplicate,
-} from './factExtractor';
-
-// Main hook for React integration
-export type {
-  UseDeepResearchOptions,
-  UseDeepResearchReturn,
-} from './useDeepResearch';
-
-export { useDeepResearch } from './useDeepResearch';
-
