+++ src/types/messages.ts
@@ -6,7 +6,6 @@
  */
 
 import type { ThreadMessageLike } from '@assistant-ui/react';
-import type { ResearchState } from '../hooks/useDeepResearch/types';
 
 /**
  * Gglib message type - directly uses ThreadMessageLike
@@ -65,10 +64,6 @@ export type GglibMessageCustom = {
   dbId?: number;
   turnId?: string;
   iteration?: number;
-  /** Deep research state - if present, this message is a research artifact */
-  researchState?: ResearchState;
-  /** Marker that this message represents a deep research session */
-  isDeepResearch?: boolean;
   /** Whether this message originated from voice input/output */
   isVoice?: boolean;
   /** Set once the final iteration is complete; triggers persisted transcript regeneration. */
